# api/openapi.yaml
openapi: 3.1.0
info:
  title: Kredit API
  version: 0.1.0
  description: Contract-first API for Kredit production v1.
servers:
  - url: /api/v1
security:
  - sessionCookie: []
paths:
...
  /auth/otp/challenges:
    post:
      security: []
      operationId: requestOTP
      summary: Request a passwordless OTP challenge
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OTPRequest'
      responses:
        '202':
          description: Challenge accepted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OTPChallengeResponse'
        '429':
          description: Rate limited or resend cooldown active.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  /auth/otp/verify:
    post:
      security: []
      operationId: verifyOTP
      summary: Verify an OTP and create a browser session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OTPVerifyRequest'
      responses:
        '200':
          description: Authenticated session created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid or expired OTP.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
...
  /mfa/totp/verify:
    post:
      operationId: verifyTOTP
      summary: Verify TOTP and elevate the current session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TOTPVerifyRequest'
      responses:
        '200':
          description: Session elevated to AAL2.
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: kredit_session
...
  schemas:
...
    Problem:
      type: object
      required: [type, title, status, detail]
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        request_id:
          type: string
    OTPRequest:
      type: object
      required: [identifier, channel]
      properties:
        identifier:
          type: string
        channel:
          type: string
          enum: [phone, email]
        purpose:
          type: string
          default: login
    OTPChallengeResponse:
      type: object
      required: [challenge_id, expires_at, channel, message]
      properties:
        challenge_id:
          type: string
        expires_at:
          type: string
          format: date-time
        channel:
          type: string
        message:
          type: string
        development_code:
          type: string
          description: Development-only value; never returned in production.
    OTPVerifyRequest:
      type: object
      required: [challenge_id, code]
      properties:
        challenge_id:
          type: string
        code:
          type: string
          minLength: 6
          maxLength: 6
        device_label:
          type: string
    TOTPVerifyRequest:
      type: object
      required: [code]
      properties:
        code:
          type: string
          minLength: 6
          maxLength: 6
    User:
      type: object
      required: [id, status, created_at]
      properties:
        id:
          type: string
        email:
          type: string
        phone:
          type: string
        display_name:
          type: string
        status:
          type: string
        created_at:
          type: string
          format: date-time
    Session:
      type: object
      required: [id, user_id, authentication_level, created_at, expires_at]
      properties:
        id:
          type: string
        user_id:
          type: string
        authentication_level:
          type: string
          enum: [AAL0, AAL1, AAL2, AAL3]
        device_label:
          type: string
        created_at:
          type: string
          format: date-time
        expires_at:
          type: string
          format: date-time
...
    AuthResponse:
      type: object
      required: [user, session]
      properties:
        user:
          $ref: '#/components/schemas/User'
        session:
          $ref: '#/components/schemas/Session'
...